Added status indicators to device card

+++ database/migrations/2024_03_20_090435_create_devices_table.php
@@ -17,11 +17,11 @@ return new class extends Migration
             $table->string('name');
             $table->string('type');
             $table->longText('description');
-            $table->float('proximity');
+            $table->integer('proximity');
             $table->string('image_url');
             $table->string('connection');
             $table->integer('connection_strength');
-            $table->float('battery');
+            $table->integer('battery');
         });
     }
